+++ Shuhari.Framework.Wpf/Controls/SimpleGridSplitter.cs
@@ -1,0 +1,15 @@
+﻿using System.Windows.Controls;
+
+namespace Shuhari.Framework.Wpf.Controls
+{
+    /// <summary>
+    /// Parent for VGridSplitter/HGridSplitter
+    /// </summary>
+    public abstract class SimpleGridSplitter : GridSplitter
+    {
+        /// <summary>
+        /// Default splitter size
+        /// </summary>
+        public const double DEFAULT_SIZE = 5.0;
+    }
+}